Fix exchange rate precision: upgrade g_usdToRate from float to double array

+++ include/app_state.h
@@ -66,13 +66,13 @@ extern const char* REFRESH_MODE_LABELS[];
 
 // ==================== Display Currency (Multi-currency) =====================
 // V0.99f: Extended to support 9 currencies
-extern int   g_displayCurrency;
-extern float g_usdToRate[CURR_COUNT];  // Exchange rates: USD -> other currencies
-extern bool  g_fxValid;
+extern int    g_displayCurrency;
+extern double g_usdToRate[CURR_COUNT];  // Exchange rates: USD -> other currencies (double for precision)
+extern bool   g_fxValid;
 extern time_t g_nextFxUpdateUtc;
 
 // Backward compatibility: g_usdToTwd is a reference to g_usdToRate[CURR_TWD]
-extern float& g_usdToTwd;
+extern double& g_usdToTwd;
 
 // ==================== Global Variables =====================
